+++ client/src/main.tsx
@@ -2,9 +2,9 @@ import ReactDOM from 'react-dom/client'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 import 'bootstrap/dist/css/bootstrap.min.css'
 
-import App from './App.jsx'
-import SearchBooks from './pages/SearchBooks'
-import SavedBooks from './pages/SavedBooks'
+import App from './App.js'
+import SearchBooks from './pages/SearchBooks.js'
+import SavedBooks from './pages/SavedBooks.js'
 
 const router = createBrowserRouter([
   {
