+++ client/src/models/Book.ts
