+++ client/src/utils/auth.ts
